requirements.txt
altair                   
argon2-cffi             
astor                
attrs                  
backcall            
backports.zoneinfo   
base58                 
bleach                   
blinker                    
cachetools              
certifi               
cffi                
charset-normalizer         
click             
cycler         
debugpy    
decorator               
defusedxml               
entrypoints                
gitdb                       
GitPython              
idna                 
ipykernel                
ipython                    
ipython-genutils      
ipywidgets                  
jedi                       
Jinja2                      
jsonschema                  
jupyter-client              
jupyter-core                
jupyterlab-pygments         
jupyterlab-widgets          
kiwisolver                  
MarkupSafe                  
matplotlib                  
matplotlib-inline           
mistune                     
nbclient                    
nbconvert                   
nbformat                   
nest-asyncio             
notebook                 
numpy                     
packaging             
pandas                     
pandocfilters             
parso                       
pexpect                     
pickleshare               
Pillow                     
pip                      
pkg-resources              
prometheus-client         
prompt-toolkit           
protobuf                   
ptyprocess                
pyarrow                     
pycparser                 
pydeck                     
Pygments                   
pyparsing                
pyrsistent               
python-dateutil             
pytz                     
pytz-deprecation-shim 
pyzmq                      
requests              
Send2Trash                 
setuptools                
six                     
smmap                      
streamlit                  
terminado                 
testpath                    
toml                     
toolz                  
tornado                      
traitlets                   
typing-extensions       
tzdata                
tzlocal                   
urllib3              
validators           
watchdog              
wcwidth                 
webencodings                
widgetsnbextension    
